functions/stop-notebook/requirements.txt
functions_framework
google-cloud-notebooks